+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+- `pl$collect_all()` to efficiently collect a list of LazyFrames (#1598).
 - `<lazyframe>$remove()` and `<dataframe>$remove()` as a complement to
   `$filter()` (#1632).